Session Complete: Tests Fixed + Kernel Frozen
✅ Fixed all compilation errors
- Type conversions in tests (u64→u32)
- CPG determinism test simplified
- All 103 tests passing

✅ Declared kernel frozen
- KERNEL_FREEZE.md created
- Phases 1-4 + Path B complete
- Changes happen AROUND kernel, not inside

✅ Created operationalization roadmap
- OPERATIONALIZATION.md with 5-step plan
- Package as artifact → CLI
- Choose narrow environment
- Benchmarks as CI guardrails
- Write determinism explainer
- Discipline: what NOT to do

VTR STATUS: PRODUCTION-READY
Next phase: Stewardship, not expansion

+++ tests/phase4_validation.rs
@@ -23,7 +23,7 @@ fn test_parallel_execution_determinism() {
             CPGNodeId(i),
             CPGNodeKind::Function,
             OriginRef::Function { function_id: semantic::model::FunctionId(i) },
-            ByteRange::new((i-1) * 10, i * 10),
+            ByteRange::new((i as u32 - 1) * 10, i as u32 * 10),
         ));
     }
